Implement MZMR logic, implement new options, rename some locations, small changes
This completes the first pass of logic, including boss events and logic-affecting options for tricks. Overall it's pretty messy right now and there are likely softlocks or odd edge cases that I missed, but I generated test seeds with different options and all seemed completable from the spoiler log!

Some locations were renamed to be a little more clear for players who may not be highly familiar with the game.

Lastly, there are some minor changes/fixes to __init__.py and items.py. Of particular note is that the Plasma Beam has been made a progression item, as it is required to defeat the Black Space Pirates guarding the ship at the end of the game.